+++ src/openvpn/buffer.h
@@ -601,26 +601,26 @@ buf_inc_len(struct buffer *buf, int inc)
  */
 
 static inline uint8_t *
-buf_prepend(struct buffer *buf, int size)
+buf_prepend(struct buffer *buf, ssize_t size)
 {
     if (!buf_valid(buf) || size < 0 || size > buf->offset)
     {
         return NULL;
     }
-    buf->offset -= size;
-    buf->len += size;
+    buf->offset -= (int)size;
+    buf->len += (int)size;
     return BPTR(buf);
 }
 
 static inline bool
-buf_advance(struct buffer *buf, int size)
+buf_advance(struct buffer *buf, ssize_t size)
 {
     if (!buf_valid(buf) || size < 0 || buf->len < size)
     {
         return false;
     }
-    buf->offset += size;
-    buf->len -= size;
+    buf->offset += (int)size;
+    buf->len -= (int)size;
     return true;
 }
 
@@ -1187,7 +1187,7 @@ struct buffer_entry *buffer_list_push_data(struct buffer_list *ol, const void *d
  */
 struct buffer *buffer_list_peek(struct buffer_list *ol);
 
-void buffer_list_advance(struct buffer_list *ol, int n);
+void buffer_list_advance(struct buffer_list *ol, ssize_t n);
 
 void buffer_list_pop(struct buffer_list *ol);
